Changing version back to 1.8
Team, if you notice that when you commit your changes that it updates this file from 1.8 to 1.7, it means your JDK is version 1.7.

For consistency's sake (and because we think it caused issues with Sue when she would attempt to push her commits, make sure you upgrade to JDK 1.8. Thanks!